Removed Repair Shop issue count from Menu. Added more home page images. Updated text on Menu.
