+++ src/pages/Repair.tsx
@@ -1,13 +1,15 @@
 import './Repair.css';
-import { GET_SONGS_WITH_ISSUES } from '../graphql/graphql';
+import { GET_SONGS_WITH_ISSUES, GET_SONGS_WITH_ISSUES_COUNT } from '../graphql/graphql';
 import ScrollingSongList from '../components/ScrollingSongList';
 import { IonButtons, IonContent, IonHeader, IonLabel, IonMenuButton, IonPage, IonSegment, IonSegmentButton, IonTitle, IonToolbar, SegmentChangeEventDetail } from '@ionic/react';
 import { useEffect, useState } from 'react';
 import SearchForSongs from '../components/SearchForSongs';
 import { Song } from '../common/types';
 import { useHistory } from 'react-router';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from '../store/store';
+import { setIssueCount } from '../store/slices/issueCountSlice';
+import { useLazyQuery } from '@apollo/client';
 
 const Repair: React.FC = () => {
 
@@ -17,6 +19,22 @@ const Repair: React.FC = () => {
 
   const [repairCount, setRepairCount] = useState<string>('0');
   const songsWithIssuesCount = useSelector((state: RootState) => state.issueCount.value);
+
+  const dispatch = useDispatch();
+
+  const [
+    getIssueCount,
+    { loading: loadingCount, error: errorCount, data: dataCount }
+  ] = useLazyQuery(GET_SONGS_WITH_ISSUES_COUNT, {
+    fetchPolicy: 'no-cache', nextFetchPolicy: 'no-cache', onCompleted: (data) => {
+      setRepairCount(data.getSongsWithIssuesCount);
+      dispatch(setIssueCount(data.getSongsWithIssuesCount));
+    },
+  });
+
+  useEffect(() => {
+    getIssueCount();
+  }, []);
 
   useEffect(() => {
     setRepairCount(songsWithIssuesCount.toString());
